added recipe section, upgraded package.json and fixed issue with swagger docs
