+++ api/babel.config.json
@@ -4,7 +4,7 @@
       "@babel/preset-env",
       {
         "useBuiltIns": "usage",
-        "corejs": "2.6.11" // match this to package.json version
+        "corejs": "3.24.1" // match this to package.json version
       }
     ]
   ]
