+++ api/package.json
@@ -21,26 +21,24 @@
   "homepage": "https://github.com/PeterSherk/vue-website#readme",
   "dependencies": {
     "bcryptjs": "2.4.3",
-    "core-js": "2.6.11",
+    "core-js": "3.24.1",
     "cors": "2.8.5",
-    "dotenv": "10.0.0",
-    "express": "4.17.1",
-    "helmet": "4.6.0",
-    "install": "^0.13.0",
+    "dotenv": "16.0.1",
+    "express": "4.18.1",
+    "helmet": "5.1.1",
     "jsonwebtoken": "8.5.1",
-    "npm": "^8.1.3",
-    "openapi-types": "9.3.1",
-    "pg": "8.7.1",
+    "openapi-types": "12.0.0",
+    "pg": "8.7.3",
     "regenerator-runtime": "0.13.9",
-    "swagger-jsdoc": "6.1.0",
-    "swagger-ui-express": "4.1.6",
-    "winston": "3.3.3"
+    "swagger-jsdoc": "6.2.4",
+    "swagger-ui-express": "4.3.0",
+    "winston": "3.8.1"
   },
   "devDependencies": {
-    "@babel/cli": "7.16.0",
-    "@babel/core": "7.16.0",
-    "@babel/node": "^7.16.0",
-    "@babel/preset-env": "7.16.0",
-    "nodemon": "2.0.15"
+    "@babel/cli": "7.18.10",
+    "@babel/core": "7.18.10",
+    "@babel/node": "7.18.10",
+    "@babel/preset-env": "7.18.10",
+    "nodemon": "2.0.19"
   }
 }
