+++ api/src/app/openapi/models/User.js
@@ -1,5 +1,5 @@
 /**
- * @swagger
+ * @openapi
  *  components:
  *    schemas:
  *      User:
